# Conflicts:
#	niffler-e-2-e-tests/src/test/java/guru/qa/niffler/db/dao/AuthUserDAO.java
#	niffler-e-2-e-tests/src/test/java/guru/qa/niffler/db/dao/AuthUserDAOHibernate.java
#	niffler-e-2-e-tests/src/test/java/guru/qa/niffler/db/dao/UserDataUserDAO.java
#	niffler-e-2-e-tests/src/test/java/guru/qa/niffler/db/dao/impl/AuthUserDAOJdbc.java
#	niffler-e-2-e-tests/src/test/java/guru/qa/niffler/db/dao/impl/AuthUserDAOSpringJdbc.java
#	niffler-e-2-e-tests/src/test/java/guru/qa/niffler/test/LoginTest.java

+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/db/dao/impl/AuthUserDAOJdbc.java
@@ -1,10 +1,15 @@
-package guru.qa.niffler.db.dao;
+package guru.qa.niffler.db.dao.impl;
 
-import guru.qa.niffler.db.DataSourceProvider;
 import guru.qa.niffler.db.ServiceDB;
+import guru.qa.niffler.db.dao.AuthUserDAO;
+import guru.qa.niffler.db.dao.UserDataUserDAO;
+import guru.qa.niffler.db.jdbc.DataSourceProvider;
 import guru.qa.niffler.db.model.Authority;
 import guru.qa.niffler.db.model.AuthorityEntity;
 import guru.qa.niffler.db.model.CurrencyValues;
+import guru.qa.niffler.db.model.auth.AuthUserEntity;
+import guru.qa.niffler.db.model.auth.Authority;
+import guru.qa.niffler.db.model.userdata.UserDataUserEntity;
 import guru.qa.niffler.db.model.UserEntity;
 
 import javax.sql.DataSource;
@@ -21,7 +26,7 @@ public class AuthUserDAOJdbc implements AuthUserDAO, UserDataUserDAO {
     private static DataSource userdataDs = DataSourceProvider.INSTANCE.getDataSource(ServiceDB.USERDATA);
 
     @Override
-    public int createUser(UserEntity user) {
+    public int createUser(AuthUserEntity user) {
         int createdRows = 0;
         try (Connection conn = authDs.getConnection()) {
 
@@ -75,7 +80,7 @@ public class AuthUserDAOJdbc implements AuthUserDAO, UserDataUserDAO {
     }
 
     @Override
-    public UserEntity updateUser(UserEntity user) {
+    public AuthUserEntity updateUser(AuthUserEntity user) {
         try (Connection conn = authDs.getConnection();
              PreparedStatement usersPs = conn.prepareStatement(
                      "UPDATE users SET " +
@@ -87,35 +92,33 @@ public class AuthUserDAOJdbc implements AuthUserDAO, UserDataUserDAO {
                              "credentials_non_expired = ? " +
                              "WHERE id = ? ")) {
 
-            usersPs.setString(1, user.getUsername());
-            usersPs.setString(2, pe.encode(user.getPassword()));
-            usersPs.setBoolean(3, user.getEnabled());
-            usersPs.setBoolean(4, user.getAccountNonExpired());
-            usersPs.setBoolean(5, user.getAccountNonLocked());
-            usersPs.setBoolean(6, user.getCredentialsNonExpired());
-            usersPs.setObject(7, user.getId());
+            usersPs.setString(2, user.getUsername());
+            usersPs.setString(3, pe.encode(user.getPassword()));
+            usersPs.setBoolean(4, user.getEnabled());
+            usersPs.setBoolean(5, user.getAccountNonExpired());
+            usersPs.setBoolean(6, user.getAccountNonLocked());
+            usersPs.setBoolean(7, user.getCredentialsNonExpired());
+            usersPs.setObject(8, user.getId());
             usersPs.executeUpdate();
+            return getUserById(user.getId());
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
-        return user;
     }
 
     @Override
-    public void deleteUserById(UUID userId) {
+    public void deleteUser(AuthUserEntity userId) {
         try (Connection conn = authDs.getConnection()) {
             conn.setAutoCommit(false);
-            try (
-                    PreparedStatement authorityPs = conn.prepareStatement(
-                            "DELETE FROM authorities WHERE user_id=?");
-                    PreparedStatement usersPs = conn.prepareStatement(
-                            "DELETE FROM users WHERE id=?")
-            ) {
+            try (PreparedStatement usersPs = conn.prepareStatement("DELETE FROM users WHERE id = ?");
+                 PreparedStatement authorityPs = conn.prepareStatement("DELETE FROM authorities WHERE user_id = ?")) {
 
-                usersPs.setObject(1, userId);
-                authorityPs.setObject(1, userId);
+                authorityPs.setObject(1, userId.getId());
+                usersPs.setObject(1, userId.getId());
+
                 authorityPs.executeUpdate();
                 usersPs.executeUpdate();
+
                 conn.commit();
                 conn.setAutoCommit(true);
             } catch (SQLException e) {
@@ -128,8 +131,8 @@ public class AuthUserDAOJdbc implements AuthUserDAO, UserDataUserDAO {
     }
 
     @Override
-    public UserEntity getUserById(UUID userId) {
-        UserEntity user = new UserEntity();
+    public AuthUserEntity getUserById(UUID userId) {
+        AuthUserEntity user = new AuthUserEntity();
         try (Connection conn = authDs.getConnection()) {
             try (PreparedStatement usersPs = conn.prepareStatement(
                     "SELECT * FROM users AS ut " +
@@ -166,7 +169,7 @@ public class AuthUserDAOJdbc implements AuthUserDAO, UserDataUserDAO {
     }
 
     @Override
-    public int createUserInUserData(UserEntity user) {
+    public int createUserInUserData(UserDataUserEntity user) {
         int createdRows;
         try (Connection conn = userdataDs.getConnection();
              PreparedStatement usersPs = conn.prepareStatement("INSERT INTO users (username, currency) VALUES (?, ?)")) {
@@ -194,13 +197,10 @@ public class AuthUserDAOJdbc implements AuthUserDAO, UserDataUserDAO {
 
     @Override
     public void deleteUserByUsernameInUserData(String username) {
-        try (Connection conn = userdataDs.getConnection()) {
-            try (PreparedStatement usersPs = conn.prepareStatement(
-                    "DELETE FROM users WHERE username = ?")) {
-
-                usersPs.setString(1, username);
-                usersPs.executeUpdate();
-            }
+        try (Connection conn = authDs.getConnection();
+             PreparedStatement usersPs = conn.prepareStatement("DELETE FROM users WHERE username = ?")) {
+            usersPs.setString(1, username);
+            usersPs.executeUpdate();
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
